Housekeeping Add log files rotation Add extra validation that a user is active
